CHG: Move wavelet middleware to `compRoutines.py`
- wavelet.py now contains pure backend code
- the new `wavelet_cf()` function is the middleware
  accepting and digesting all frontend related parameters

Changes to be committed:
	modified:   ../../devtest_frontend.py
	modified:   compRoutines.py
	modified:   freqanalysis.py
	modified:   superlet.py
	modified:   wavelet.py
