+++ devtest_frontend.py
@@ -12,21 +12,22 @@ r_mtm = freqanalysis(tdat)
 
 toi_ival = np.linspace(-0.5, 1, 100)
 
-#toi_ival = [0,0.2,0.5,1]
+# toi_ival = [0,0.2,0.5,1]
 toi_ival = 'all'
-foi = np.logspace(-1, 2.6, 25)
+foi = np.logspace(-1, 2.6, 55)
 # test classical wavelet analysis
 r_wav = freqanalysis(tdat, method="wavelet",
                      toi=toi_ival,
                      output='abs',
-                     foi=None) #, foilim=[5, 500])
+                     foi=foi,
+                     # foilim=[5, 500]
+                     )
 
 # test superlet analysis
 r_sup = freqanalysis(tdat, method="superlet", toi=toi_ival,
                      order_max=20, output='abs',
                      order_min=1,
-                     c_1 = 5,
+                     c_1 = 3,
+                     # foilim=[5, 500],
+                     foi=foi,                     
                      adaptive=True)
-
-r_sup = freqanalysis(tdat, method="superlet", toi='all', order_max=30, foi=foi, output='abs',order_min=5, adaptive=True)
-#res_strials = [t for t in r_sup.trials]
